feat: テナント管理機能の拡張とエラーハンドリングの改善
- テナントレスポンスにオーナーIDを追加しました。
- テナントの作成、更新、取得、削除時にオーナーIDを処理するようにしました。
- PostgreSQLのテナントリポジトリにオーナーIDを追加し、関連するSQLクエリを更新しました。
- テナントエンティティにオーナーIDを追加しました。
- テナント作成および更新ユースケースでイベントパブリッシャーを統合しました。
- バルクエラー応答を追加し、エラーハンドリングを改善しました。
- ワークフロー関連のエラーハンドリングを改善し、エラーレスポンスを統一しました。
- ワークフローのレスポンスにステップ情報を追加しました。

+++ regions/system/server/rust/ratelimit/src/domain/entity/rate_limit_rule.rs
@@ -42,8 +42,8 @@ pub struct RateLimitRule {
     pub id: Uuid,
     pub scope: String,
     pub identifier_pattern: String,
-    pub limit: i64,
-    pub window_seconds: i64,
+    pub limit: u32,
+    pub window_seconds: u32,
     pub algorithm: Algorithm,
     pub enabled: bool,
     pub created_at: DateTime<Utc>,
@@ -51,7 +51,13 @@ pub struct RateLimitRule {
 }
 
 impl RateLimitRule {
-    pub fn new(scope: String, identifier_pattern: String, limit: i64, window_seconds: i64, algorithm: Algorithm) -> Self {
+    pub fn new(
+        scope: String,
+        identifier_pattern: String,
+        limit: u32,
+        window_seconds: u32,
+        algorithm: Algorithm,
+    ) -> Self {
         let now = Utc::now();
         Self {
             id: Uuid::new_v4(),
